Create README.md
I have one year of experience as a stock controller at a computer warehouse. I try to spend 2 hours a day learning front end and 2 hours on CCST and/or CompTia A+

I am willing to learn and put in the time required to achieve my goals. I am currently looking a help desk learnership.